Removing contact form
