--- view/adminhtml/layout/algolia_algoliasearch_support_contact.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <head>
-        <css src="Algolia_AlgoliaSearch::css/support.css"/>
-        <script src="Algolia_AlgoliaSearch::js/support.js"/>
-    </head>
-    <body>
-        <referenceContainer name="content">
-            <block class="Magento\Backend\Block\Template" name="support_contact" template="Algolia_AlgoliaSearch::support/contact.phtml" >
-                <arguments>
-                    <argument name="view_model" xsi:type="object">Algolia\AlgoliaSearch\ViewModel\Adminhtml\Support\Contact</argument>
-                </arguments>
-            </block>
-        </referenceContainer>
-    </body>
-</page>
