+++ view/adminhtml/templates/support/overview.phtml
@@ -25,11 +25,6 @@ $videoFeatures = $block->getViewFileUrl('Algolia_AlgoliaSearch::images/video-fea
                 <span><?php echo $block->escapeHtml(__('Search Documentation')) ?></span>
             </a>
         </li>
-        <li role="tab" aria-controls="algolia-contact-panel" aria-selected="false">
-            <a href="javascript:;">
-                <span><?php echo $block->escapeHtml(__('Contact Us')) ?></span>
-            </a>
-        </li>
     </ul>
 </div>
 
@@ -250,9 +245,4 @@ $videoFeatures = $block->getViewFileUrl('Algolia_AlgoliaSearch::images/video-fea
             </div>
         </div>
     </div>
-
-    <div class="algolia_support_panel" id="algolia-contact-panel">
-        <?php echo $block->getChildHtml('support_contact') ?>
-    </div>
-
 </div>
